requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
elasticsearch==7.0.2
Flask==1.0.3
flask-mwoauth==0.3.70
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mwoauth==0.3.3
mysql-connector-python==8.0.16
oauthlib==3.0.1
protobuf==3.7.1
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
